\name{varEst-package}
\alias{varEst-package}
\alias{varEst}
\docType{package}
\title{
\packageTitle{varEst}
}
\description{
\packageDescription{varEst}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{varEst}
\packageIndices{varEst}

}
\author{
\packageAuthor{varEst}

Maintainer: \packageMaintainer{varEst}
}
\references{
Fan, J., Guo, S., Hao, N. (2012).Variance estimation using refitted cross-validation in ultrahigh dimensional regression. \emph{Journal of the Royal Statistical Society}, 74(1), 37-65\cr Ravikumar, P., Lafferty, J., Liu, H. and Wasserman, L. (2009). Sparse additive models. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 71(5), 1009-1030\cr Tibshirani, R. (1996). Regression shrinkage and selection via the Lasso. \emph{Journal of Royal Statistical Society}, 58, 267-288
}


